improvements to tags and bug fixes
Fixed bug with update_queue when there were no files in the directory.

Improved support for tags. They can now be added in the admin panel per project.

Added support for filtering of tags on results pages.

Fixed bug with the project list page when a project does not yet have a searchsetting.

Tweaked some display names for filters.